feat: add vertical-size setting to Quake Terminal extension
The Quake Terminal extension now includes a new setting called "vertical-size" which determines how
far down the terminal will drop when toggled. The default value is set to 50, but can be adjusted
between 25 and 100. This allows users to customize the size of the terminal window to their preference.

This commit also includes some minor code refactoring and bug fixes.