Allow interrupting the sliding animation
Up until now, one has to wait for the sliding animation to complete
before the hotkey has an affect again.

With a configured long animation time, that forces users to wait for
some time to correct for example an accidental hit of the hotkey.

Removing the in-transition tracker speeds up usage, ups felt
performance and reduces code complexity.