Remove always-on-top
Prior to this commit, the terminal window has been configured to always
stay on top of other applications.

Given the terminal is configured to stay visible even when focus is
lost, situations arise where newly opened apps are displayed
behind the terminal and thus, are not accessible by mouse.

Disabling always-on-top allows for new apps to overlap the terminal
window and thus, use them immediately. Hitting the hotkey brings the
terminal back to the front to be used as well.

When autohide is active, the terminal hides itself anyways when
focus is lost, so always-on-top does not make sense there anyways.