feat: add "Always On Top" option to Quake Terminal preferences
This commit adds a new option to the Quake Terminal preferences called "Always On Top".
When enabled, the terminal window will appear on top of all other non-topmost windows.
This can be useful for keeping the terminal easily accessible and visible at all times.

The new option is added to the preferences UI and is bound to the corresponding setting
in the `prefs.js` file. Additionally, the `quake-mode.js` file is updated to handle the
changes in the "Always On Top" setting. When the setting is enabled, the terminal window
is made above other windows, and when it is disabled, the terminal window is unmade above.

This commit also updates the default value of the "Auto Hide Terminal" setting in the
`org.gnome.shell.extensions.quake-terminal.gschema.xml` file to `true`, indicating that
the terminal window should be automatically hidden when it loses focus.

Overall, these changes enhance the functionality of Quake Terminal by providing more control
over the visibility and positioning of the terminal window.