feat: Add _terminalWindow property to QuakeMode class
This commit adds a new `_terminalWindow` property to the `QuakeMode` class in the `quake-mode.js` file.

This property is used to store the reference to the terminal window object.

Previously, the terminal window object was retrieved from the terminal
using the `get_windows()` method every time it was needed. Now, the
terminal window object is stored in the `_terminalWindow` property and
only retrieved once, improving performance.

This change also includes some code formatting improvements and minor bug fixes.

Closes #27